deleted/split authorization
deleted unnecessary code from authorization split up authorization into different parts with login being in different file
