+++ src/authorisation.js
@@ -1,17 +1,10 @@
-/**
- * This is an example of a basic node.js script that performs
- * the Authorization Code with PKCE oAuth2 flow to authenticate 
- * against the Spotify Accounts.
- *
- * For more information, read
- * https://developer.spotify.com/documentation/web-api/tutorials/code-pkce-flow
- */
-
+//This file is for authorisation of logins and api calls
 const clientId = '03ce66bd50a54bb78f5d89c1c401407e'; // your clientId
 const redirectUrl = 'http://localhost:3000';        // your redirect URL - must be localhost URL and/or HTTPS
 
 const authorizationEndpoint = "https://accounts.spotify.com/authorize";
 const tokenEndpoint = "https://accounts.spotify.com/api/token";
+//need to add scope for search and others that need to be done
 const scope = 'user-read-private user-read-email';
 
 // Data structure that manages the current active token, caching it in localStorage
@@ -32,35 +25,6 @@ const currentToken = {
     localStorage.setItem('expires', expiry);
   }
 };
-
-// On page load, try to fetch auth code from current browser search URL
-const args = new URLSearchParams(window.location.search);
-const code = args.get('code');
-
-// If we find a code, we're in a callback, do a token exchange
-if (code) {
-  const token = await getToken(code);
-  currentToken.save(token);
-
-  // Remove code from URL so we can refresh correctly.
-  const url = new URL(window.location.href);
-  url.searchParams.delete("code");
-
-  const updatedUrl = url.search ? url.href : url.href.replace('?', '');
-  window.history.replaceState({}, document.title, updatedUrl);
-}
-
-// If we have a token, we're logged in, so fetch user data and render logged in template
-if (currentToken.access_token) {
-  const userData = await getUserData();
-  //renderTemplate("main", "logged-in-template", userData);
-  //renderTemplate("oauth", "oauth-template", currentToken);
-}
-
-// Otherwise we're not logged in, so render the login template
-//if (!currentToken.access_token) {
-  //renderTemplate("main", "login");
-//}
 
 async function redirectToSpotifyAuthorize() {
   const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
@@ -151,42 +115,7 @@ async function logoutClick() {
 async function refreshTokenClick() {
   const token = await refreshToken();
   currentToken.save(token);
-  //renderTemplate("oauth", "oauth-template", currentToken);
 }
 
 
-
-
-
-//Not sure if this is right will have to change in order to fit for my app.js NOT index.js
-// HTML Template Rendering with basic data binding - demoware only.
-/*function renderTemplate(targetId, templateId, data = null) {
-  const template = document.getElementById(templateId);
-  const clone = template.content.cloneNode(true);
-
-  const elements = clone.querySelectorAll("*");
-  elements.forEach(ele => {
-    const bindingAttrs = [...ele.attributes].filter(a => a.name.startsWith("data-bind"));
-
-    bindingAttrs.forEach(attr => {
-      const target = attr.name.replace(/data-bind-/, "").replace(/data-bind/, "");
-      const targetType = target.startsWith("onclick") ? "HANDLER" : "PROPERTY";
-      const targetProp = target === "" ? "innerHTML" : target;
-
-      const prefix = targetType === "PROPERTY" ? "data." : "";
-      const expression = prefix + attr.value.replace(/;\n\r\n/g, "");
-
-      // Maybe use a framework with more validation here ;)
-      try {
-        ele[targetProp] = targetType === "PROPERTY" ? eval(expression) : () => { eval(expression) };
-        ele.removeAttribute(attr.name);
-      } catch (ex) {
-        console.error(`Error binding ${expression} to ${targetProp}`, ex);
-      }
-    });
-  });
-
-  const target = document.getElementById(targetId);
-  target.innerHTML = "";
-  target.appendChild(clone);
-}*/
+export { redirectToSpotifyAuthorize, getToken, currentToken };
